+++ src/main/java/com/myspringdemo/blog/repo/RolesRepository.java
@@ -4,8 +4,6 @@ import com.myspringdemo.blog.models.ERole;
 import com.myspringdemo.blog.models.Role;
 import org.springframework.data.repository.CrudRepository;
 
-import java.util.List;
-
-public interface RolesRepository extends CrudRepository<Role, Integer> {
+public interface RolesRepository extends CrudRepository<Role, Long> {
     Role findByName(ERole name);
 }
